# openapi.yaml
openapi: 3.0.3
info:
  title: Clothes E-commerce API
  version: 0.1.0
servers:
  - url: http://localhost:3000
paths:
...
  /api/register:
    post:
      summary: Register a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string }
                password: { type: string }
                name: { type: string }
      responses:
        '200': { description: User created }
        '400': { description: Invalid input }
...